Stop steady state detection after it was detected
As there is little point in continuing the search for
something that was already found, steady state detection
no longer takes place after steady state was detected.

If a packet or ack is lost after a previously detected
steady state, history is erased and the search is on for
a new steady state!