BUGFIX: Fixed NullPointerException in FlowForm
While executing FlowForm's InitializeComponent, the ClientSizeChanged
event is fired. The handler for that event relies on an existing
m_bitmap member. This member, however, only gets initialized further
down FlowForm's constructor (after the InitializeComponent call),
leading to a NullPointerException in the ClientSizeChanged event
handler. This has been fixed by only attaching the handler to the
event at the end of FlowForm's constructor.