BUGFIX? recurring history is not always steady
In some cases, recurring history does not reliably indicate
steady state.

This has been mitigated by demanding the steady state period
to be at least as long as the network delay, but it is an ad
hoc fix and not really based on solid reasoning...

It solves the following test case though:
For a buffer size of 8 packets with congestion control
disabled, steady state is detected when the receiver buffer
contains one packet three receiver-state-updates in a row.

Thought: would this be solved if we demanded there was at
least one round-trip (an actual round trip, as in:
"a correctly ACK'ed packet", not just twice the netork delay)
in the steady state region? I'm not sure... This might well
lead to the detected steady state region containing a
multiple of the actual steady state...